Rename provider module
